requirements.txt
flask
requests
faker
python-dotenv
